Update patch set 2
Patch Set 2: Code-Review+1

Reviewing, I understand that this logging reference will likely change as the rest of the RI code comes in.  As agreed we will modify as it is implemented in the spring framework.  I also need to correlate this example with the wiki guidelines and see it in action after we push the docker/k8s infrastructure

Patch-set: 2