Update patch set 1
Patch Set 1:

I understand that setting a default elapsed time can be confusing. The initial logic behind this was to keep it so that the log line would still be valid if elapsedTime could not be calculated. I am seeing it where entryTimestamp is null so a NullPointerException is thrown. However, since the entry or invoke timestamps will also be null in that case, that would make the log lines invalid anyway. I can remove setting the default elapsedTime, but I think that we should still catch exceptions and print a warning to identify that we weren't able to set elapsedTime

Patch-set: 1