Update patch set 10
Patch Set 10:

Updated the JIRA and commit message after carefully re-reviewing the spec.

The new log format looks like this if it can't determine the value of partner name
Checking Authorization header to determine the value of PartnerName
Checking X-ONAP-PartnerName header to determine the value of PartnerName
Checking User-Agent header to determine the value of PartnerName
Checking X-ClientID header to determine the value of PartnerName
PartnerName value could not be determined, defaulting partnerName to UNKNOWN.

You can run the junit locally to see the same.

Patch-set: 10