Update patch set 1
Patch Set 1:

I added changes here to split InvocationId into two MDC values. These changes will break SO unit tests and will affect other functionality as there is some getting/setting of InvocationID in other aspects of the code that need to be changed or updated. I changed the SNAPSHOT version here so that we could keep the 1.6.2-SNAPSHOT version for now until I'm able to update SO to point to this new version, and add changes to fix those issues. Please let me know if there is a better way to handle this version issue.

Patch-set: 1